Troubleshooting with the pivot motors
Redid the angular vector calculation, commented out translation to further study

marc if you are reading this I accidentally clones the FR translation which is why the BL wasn't moving, oops

+++ Milan/src/main/include/commands/baseControl.h
@@ -12,9 +12,9 @@
 class baseControl : public frc::Command {
  public:
  
-  int leftStickX;
-  int leftStickY;
-  int rightStickX;
+  double leftStickX;
+  double leftStickY;
+  double rightStickX;
 
   baseControl();
   void Initialize() override;
